wrote quadrature into existing ARTS file
