+++ konrad/radiation/__init__.py
@@ -25,7 +25,7 @@ and heating rates:
 from .arts import ARTS
 from .radiation import Radiation
 from .rrtmg import RRTMG
-from .quadrature import quadrature
+#from .quadrature import quadrature
 from .common import *
