+++ konrad/radiation/arts.py
@@ -13,12 +13,13 @@ from konrad.cloud import ClearSky
 from .rrtmg import RRTMG
 from .common import fluxes2heating
 
+from scipy.constants import speed_of_light as c
 
 logger = logging.getLogger(__name__)
 
 
 class _ARTS:
-    def __init__(self, ws=None, threads=None, nstreams=4, scale_vmr=True, verbosity=0):
+    def __init__(self, ws=None, threads=None, nstreams=4, scale_vmr=True, verbosity=0, quadrature = False, quadrature_filename=join(dirname(__file__), "data", '64_quadrature.h5'), lookup_filename=join(dirname(__file__), "data", 'quadrature_lookup.xml')):
         """Initialize a wrapper for an ARTS workspace.
 
         Parameters:
@@ -30,11 +31,23 @@ class _ARTS:
             scale_vmr (bool): Control whether dry volume mixing ratios are
                 scaled with the water-vapor concentration (default is `False.`)
             verbosity (int): Control the ARTS verbosity from 0 (quiet) to 2.
+            
+            quadrature (bool): use quadrature scheme or LBL calculation via ARTS (default)
+            quadrature_filename (str): where to find the file that contains the wavenumbers (cm^{-1})
+                and weights to compute the quadrature scheme. This file must contain at least the variable S (wavenumbers) and
+            W (weights). Default is the result of a quadrature scheme trained on 72,000 frequencies x 50 profiles x 55 vertical
+                levels generated with pyARTS spectroscopy with CKDMIP's eval1 atmospheric conditions, resulting in 64
+            frequencies/weights, found in the data folder. The wavenumbers should be in ascending order and the weights
+                should be in the corresponding order.
+            lookup_filename (str): filename where the lookup table corresponding to the chosen frequencies can be found. If
+                None, use ARTS to generate monochromatic fluxes on the fly.
+
         """
         from pyarts.workspace import Workspace
 
         self.nstreams = nstreams
         self.scale_vmr = scale_vmr
+        self._quadrature = quadrature
 
         if ws is None:
             self.ws = Workspace(verbosity=verbosity)
@@ -72,33 +85,103 @@ class _ARTS:
                 "CO",
             ]
         )
+        
+        if (self._quadrature == True):
+            import xarray as xr
+            import pyarts
 
-        # Read lookup table
-        abs_lookup = os.getenv(
-            "KONRAD_LOOKUP_TABLE", join(dirname(__file__), "data/abs_lookup.xml")
-        )
+            ### use quadrature scheme
+            self._lookup_filename = lookup_filename
 
-        self.ws.abs_lookup_is_adapted.initialize_if_not()
-        if isfile(abs_lookup):
-            self.ws.abs_lines_per_speciesSetEmpty()
-            self.ws.ReadXML(self.ws.abs_lookup, abs_lookup)
-            self.ws.f_gridFromGasAbsLookup()
-            self.ws.abs_lookupAdapt()
-            self.ws.propmat_clearsky_agendaAuto(use_abs_lookup=1)
+            # requires os.environ['ARTS_DATA_PATH'] to be set to the location of arts-cat-data
+            line_basename="lines/"
+            xsec_basename="xsec/"
+            cia_basename="cia-xml/hitran2011/hitran_cia2012_adapted.xml.gz"
 
-            self.ws.jacobianOff()  # No jacobian calculation
-            self.ws.cloudboxOff()  # Clearsky = No scattering
-            self.ws.sensorOff()  # No sensor properties
+            # No jacobian calculation
+            self.ws.jacobianOff()
+
+            # Clearsky = No scattering
+            self.ws.cloudboxOff()
+
+            # Output radiance not converted
+            self.ws.StringSet(self.ws.iy_unit, "1")
+
+            # set particle scattering to zero, because we want only clear sky
+            self.ws.scat_data_checked = 1
+            self.ws.Touch(self.ws.scat_data)
+
+            # open quadrature/weight combo
+            self._optimization_result = xr.open_dataset(quadrature_filename, engine = "netcdf4")
+
+            self._S = self._optimization_result.S.data
+            self._W = self._optimization_result.W.data
+
+            self.ws.f_grid = pyarts.arts.convert.kaycm2freq(self._S)
+
+            if (lookup_filename == None):
+                self.ws.ReadXML(
+                    self.ws.predefined_model_data,
+                    "model/mt_ckd_4.0/H2O.xml"
+                )
+
+                # Read line catalog
+                self.ws.abs_lines_per_speciesReadSpeciesSplitCatalog(basename=line_basename)
+
+                # Read cross section data
+                self.ws.ReadXsecData(basename=xsec_basename)
+
+                # Read CIA data
+                self.ws.abs_cia_dataReadFromXML(filename = cia_basename)
+
+                # Set line shape and cut off.
+                self.ws.abs_lines_per_speciesCompact()  # Throw away lines outside f_grid
+                self.ws.abs_lines_per_speciesLineShapeType(self.ws.abs_lines_per_species, "VP")
+                self.ws.abs_lines_per_speciesNormalization(self.ws.abs_lines_per_species, "VVH")
+                self.ws.abs_lines_per_speciesCutoff(self.ws.abs_lines_per_species, "ByLine", 750e9)
+                self.ws.abs_lines_per_speciesTurnOffLineMixing()
+                self.ws.propmat_clearsky_agendaAuto(T_extrapolfac = 1E99)
+
+            else:
+                abs_lookup = lookup_filename
+                self.ws.abs_lookup_is_adapted.initialize_if_not()
+
+                self.ws.abs_lines_per_speciesSetEmpty()
+                self.ws.ReadXML(self.ws.abs_lookup, abs_lookup)
+                self.ws.f_gridFromGasAbsLookup()
+                self.ws.abs_lookupAdapt()
+                self.ws.propmat_clearsky_agendaAuto(use_abs_lookup=1)
+
+                self.ws.sensorOff()  # No sensor properties
+            
         else:
-            warnings.warn(
-                "Could not find ARTS absorption lookup table.\n"
-                "To perform ARTS calculations you have to download the lookup "
-                "table at:\n\n    https://doi.org/10.5281/zenodo.3885410\n\n"
-                "Afterwards, use the following environment variable to tell "
-                "konrad where to find it:\n\n"
-                "    $ export KONRAD_LOOKUP_TABLE='/path/to/abs_lookup.xml'",
-                UserWarning,
+            ### use ARTS LBL with lookup table
+            # Read lookup table
+            abs_lookup = os.getenv(
+                "KONRAD_LOOKUP_TABLE", join(dirname(__file__), "data/abs_lookup.xml")
             )
+
+            self.ws.abs_lookup_is_adapted.initialize_if_not()
+            if isfile(abs_lookup):
+                self.ws.abs_lines_per_speciesSetEmpty()
+                self.ws.ReadXML(self.ws.abs_lookup, abs_lookup)
+                self.ws.f_gridFromGasAbsLookup()
+                self.ws.abs_lookupAdapt()
+                self.ws.propmat_clearsky_agendaAuto(use_abs_lookup=1)
+
+                self.ws.jacobianOff()  # No jacobian calculation
+                self.ws.cloudboxOff()  # Clearsky = No scattering
+                self.ws.sensorOff()  # No sensor properties
+            else:
+                warnings.warn(
+                    "Could not find ARTS absorption lookup table.\n"
+                    "To perform ARTS calculations you have to download the lookup "
+                    "table at:\n\n    https://doi.org/10.5281/zenodo.3885410\n\n"
+                    "Afterwards, use the following environment variable to tell "
+                    "konrad where to find it:\n\n"
+                    "    $ export KONRAD_LOOKUP_TABLE='/path/to/abs_lookup.xml'",
+                    UserWarning,
+                )
 
         # Set number of OMP threads
         if threads is not None:
@@ -240,12 +323,56 @@ class _ARTS:
         self.ws.t_surface = np.array([[t_surface]])
         self.ws.z_surface = np.array([[0.0]])
         self.ws.z_field.value[0, 0, 0] = 0.0
+        self.ws.surface_skin_t = self.ws.t_field.value[0, 0, 0]
+        
+        # set cloudbox to full atmosphere
+        self.ws.cloudboxSetFullAtm()
+        self.ws.pnd_fieldZero()
 
         # Perform configuration and atmosphere checks
         self.ws.atmfields_checkedCalc()
         self.ws.propmat_clearsky_agenda_checkedCalc()
         self.ws.atmgeom_checkedCalc()
         self.ws.cloudbox_checkedCalc()
+        
+    def calc_monochromatic_fluxes(self, atmosphere, t_surface):
+        """Calculate the spectral irradiance field."""
+
+        self.set_atmospheric_state(atmosphere, t_surface)
+        
+        if self._lookup_filename == None:
+
+            self.ws.scat_data_checkedCalc()
+            self.ws.lbl_checkedCalc()
+
+        # get the zenith angle grid and the integrations weights
+        self.ws.AngularGridsSetFluxCalc(
+            N_za_grid=self.nstreams, N_aa_grid=1, za_grid_type="double_gauss"
+        )
+
+        # calculate intensity field
+        self.ws.Tensor3Create("trans_field")
+
+        self.ws.spectral_radiance_fieldClearskyPlaneParallel(
+            trans_field=self.ws.trans_field,
+            use_parallel_za=0,
+        )
+        self.ws.spectral_irradiance_fieldFromSpectralRadianceField()
+
+        # Perform RT calculations
+        #self.ws.DisortCalcIrradiance(nstreams=self.nstreams, emission=1)
+        
+        spec_flux_up = self.ws.spectral_irradiance_field.value[:, :, 0, 0, 1]
+        spec_flux_down = self.ws.spectral_irradiance_field.value[:, :, 0, 0, 0]
+        spec_flux_up = (spec_flux_up * (c * 100))
+        spec_flux_down = (spec_flux_down * (c * 100))
+
+
+        return (
+            spec_flux_up, -spec_flux_down
+            
+        )
+
 
     def calc_spectral_irradiance_field(self, atmosphere, t_surface):
         """Calculate the spectral irradiance field."""
@@ -287,6 +414,10 @@ class _ARTS:
         )
 
         return self.ws.f_grid.value[:].copy(), tau
+    
+    def integrate_quadrature_irradiance(self, spectral_flux):
+        int_flux = np.matmul(self._W, spectral_flux)
+        return int_flux
 
     @staticmethod
     def integrate_spectral_irradiance(frequency, irradiance):
@@ -357,11 +488,21 @@ class ARTS(RRTMG):
         )
         sw_fluxes = sw_dT_fluxes[1]
 
-        # Perform ARTS simulation
-        f, _, irradiance_field, _ = self._arts.calc_spectral_irradiance_field(
-            atmosphere=atmosphere, t_surface=surface["temperature"][0]
-        )
-        Fd, Fu = self._arts.integrate_spectral_irradiance(f, irradiance_field)
+        if (self._arts._quadrature == True):
+            # Perform quadrature simulation
+            spec_flux_up, spec_flux_down = self._arts.calc_monochromatic_fluxes(
+                atmosphere=atmosphere, t_surface=surface["temperature"][0]
+            )
+
+            Fd = self._arts.integrate_quadrature_irradiance(spectral_flux = spec_flux_down)
+            Fu = self._arts.integrate_quadrature_irradiance(spectral_flux = spec_flux_up)
+
+        else:
+            # Perform ARTS simulation
+            f, _, irradiance_field, _ = self._arts.calc_spectral_irradiance_field(
+                atmosphere=atmosphere, t_surface=surface["temperature"][0]
+            )
+            Fd, Fu = self._arts.integrate_spectral_irradiance(f, irradiance_field)
 
         # Interpolate RT results on fine original grid
         def _reshape(x, trim=-1):
